Rename route package to api
