+++ app/api/api/healthcheck.py
@@ -7,8 +7,8 @@ from sqlalchemy import text
 from werkzeug.exceptions import ServiceUnavailable
 
 import api.adapters.db.flask_db as flask_db
-from api.route import response
-from api.route.schemas import request_schema
+from api.api import response
+from api.api.schemas import request_schema
 
 logger = logging.getLogger(__name__)
