+++ app/api/app.py
@@ -11,9 +11,9 @@ import api.adapters.db.flask_db as flask_db
 import api.logging
 import api.logging.flask_logger as flask_logger
 from api.auth.api_key_auth import User, get_app_security_scheme
-from api.route.healthcheck import healthcheck_blueprint
-from api.route.schemas import response_schema
-from api.route.user_route import user_blueprint
+from api.api.healthcheck import healthcheck_blueprint
+from api.api.schemas import response_schema
+from api.api.users import user_blueprint
 
 logger = logging.getLogger(__name__)
